navigation/requirements.txt
python-dotenv==0.19.1
RPi.GPIO==0.7.0
